Update project for version 1.1.6 release
This commit includes updates to several files in preparation for the version 1.1.6 release. The changes include updates to the Redstone and Rednet blocks, improvements to the Lua generator, and updates to the project dependencies in package.json. The UPDATE.md file has also been updated with the latest release notes.